fix: 심판 클래스 추가 - 값을 비교 판정하고 힌트를 제공하는 기능을 심판 클래스로 분리하여 구현
