+++ README.md
@@ -16,14 +16,15 @@
 3. 게임플레이어(유저) 객체
 4. 비정상적 입력을 판별하는 객체
 5. 투구(값) 객체
+6. 심판 객체
 
 ## 각 객체의 속성(변수)과 기능(메서드) 판단하기
 1. 
 속성: 게임 코인
 기능: 게임 시작시키기, 게임 코인 여부 판단하기
 2.
-속성: 해당 게임의 정답 값, 정답과 추측의 비교 판정 값
-기능: 정답 값 선택하기, 정답과 추측 비교 판정하기
+속성: 해당 게임의 정답 값
+기능: 정답 값 선택하기
 3.
 속성: 추측 값
 기능: 추측 값 선택하기
@@ -34,6 +35,10 @@
 5.
 속성: 1~9 세자리로 구성된 숫자
 기능: 필요 x
+
+6.
+속성: 정답과 추측의 비교 판정 값
+기능: 정답과 추측 비교 판정하기
 
 ## 구현할 기능
 1. 게임 코인이 들어오는지 판단할 기능
