+++ src/main/java/baseball/GameComputer.java
@@ -40,7 +40,7 @@ public class GameComputer {
     static boolean startGame(String computerChoice) {
         GamePlayer player = new GamePlayer();
         player.chooseChoice();
-        int strike = umpirePitch(computerChoice, player.userChoice.choice);
+        int strike = GameUmpire.umpirePitch(computerChoice, player.userChoice.choice);
         if (strike == BallChoice.CHOICE_SIZE) {
             System.out.println("3개의 숫자를 모두 맞히셨습니다! 게임 종료");
             return true;
@@ -49,46 +49,4 @@ public class GameComputer {
         }
     }
 
-    static int umpirePitch(String computerChoice, String userChoice) {
-        int strike = 0;
-        int ball = 0;
-        for (int i = 0; i < computerChoice.length(); i++) {
-            if (computerChoice.charAt(i) == userChoice.charAt(i)) {
-                strike += 1;
-            }
-        }
-
-        for (int i = 0; i < userChoice.length(); i++) {
-            if (calculateBall(computerChoice, userChoice.charAt(i))) {
-                ball += 1;
-            }
-        }
-
-        ball -= strike;
-        callUmpire(strike,ball);
-        return strike;
-    }
-
-    static boolean calculateBall(String computerChoice, int pitch) {
-        boolean ball = false;
-
-        for (int i = 0; i<computerChoice.length(); i++) {
-            if (computerChoice.charAt(i) == pitch) {
-                ball = true;
-            }
-        }
-        return ball;
-    }
-
-    static void callUmpire(int strike, int ball) {
-        if (strike == 0 && ball == 0) {
-            System.out.println("낫싱");
-        }else if (strike == 0 && ball != 0) {
-            System.out.println(ball+"볼");
-        }else if (strike !=0 && ball == 0) {
-            System.out.println(strike+"스트라이크");
-        }else {
-            System.out.println(ball+"볼 "+strike+"스트라이크");
-        }
-    }
 }
